+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/ToolBar/ToolBarBottom/ToolBarBottomConfigurationView.swift
@@ -102,6 +102,12 @@ struct ToolBarBottomConfiguration: View {
                     ToolBarTrailingConfiguration(configurationModel: configurationModel)
                 }
 
+                if !configurationModel.groupedItems, configurationModel.trailing == .icon {
+                    OUDSChipPicker(title: "app_components_badge_tech".localized(),
+                                   selection: $configurationModel.badgeType,
+                                   chips: BarItemBadgeType.chips)
+                }
+
                 ToolBarItemStyle(configurationModel: configurationModel)
 
                 DesignToolboxEditContentDisclosure {