+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/ToolBar/ToolBarBottom/ToolBarBottomPage.swift
@@ -49,7 +49,7 @@ private struct ToolBarBottomDemo: View {
     @Environment(\.theme) private var theme
 
     var body: some View {
-        OUDSButton(text: "app_components_toolbar_demo_showDemo_label", appearance: .strong) {
+        OUDSButton("app_components_toolbar_demo_showDemo_label", appearance: .strong) {
             configurationModel.showDemo = true
         }
         .disabled(configurationModel.leading == .none && configurationModel.trailing == .none)
@@ -86,7 +86,7 @@ private struct BottomBarModifier: ViewModifier {
                 .toolBarBottom {
                     configurationModel.leadingItems(for: theme)
                 } trailingItems: {
-                    configurationModel.trailingItems(for: theme)
+                    configurationModel.trailingItems()
                 }
         }
     }