+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/ToolBar/ToolBarTop/ToolBarTopPage.swift
@@ -50,7 +50,7 @@ private struct ToolBarTopDemo: View {
 
     var body: some View {
         NavigationLink(destination: demo(), isActive: $configurationModel.showNavigation) {
-            OUDSButton(text: "app_components_toolbar_demo_showDemo_label", appearance: .strong) {
+            OUDSButton("app_components_toolbar_demo_showDemo_label", appearance: .strong) {
                 switch configurationModel.demoOption {
                 case .navigation:
                     configurationModel.showNavigation = true
@@ -82,7 +82,7 @@ private struct ToolBarTopDemo: View {
             {
                 configurationModel.leadingItems(for: theme)
             } trailingItems: {
-                configurationModel.trailingItems(for: theme)
+                configurationModel.trailingItems()
             }
     }